requirements.txt
astroid==1.5.3
cycler==0.10.0
isort==4.2.15
jieba==0.39
langid==1.1.6
lazy-object-proxy==1.3.1
matplotlib==2.1.0
mccabe==0.6.1
memory-profiler==0.47
numpy==1.13.3
psutil==5.4.0
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
six==1.11.0
slugid==1.0.7
tqdm==4.19.4
warc3-wet==0.2.3
wrapt==1.10.11
